move remaining functions from core to pkg/visitor
