+++ pkg/visitor/verify.go
@@ -1,4 +1,4 @@
-package core
+package visitor
 
 import (
 	"context"
